// Assets-Uncapped/front-portal-develop/src/services/api/kyc/models/CustomerPersonDTO.ts
export interface CustomerPersonDTO {
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    firstName?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    lastName?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    phone?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    organisationId?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    status?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    externalId: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    verificationLink?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    ssn?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    passportNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    dateOfBirth?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    nationality?: string;
    /**
     * 
     * @type {CustomerPersonAddressDTO}
     * @memberof CustomerPersonDTO
     */
    address?: CustomerPersonAddressDTO;
    /**
     * 
     * @type {CustomerPersonAddressDTO}
     * @memberof CustomerPersonDTO
     */
    deliveryAddress?: CustomerPersonAddressDTO;
    /**
     * 
     * @type {string}
     * @memberof CustomerPersonDTO
     */
    livenessCheckStatus?: CustomerPersonDTOLivenessCheckStatusEnum;
}


/**
 * @export
 */
export const CustomerPersonDTOLivenessCheckStatusEnum = {
    Unverified: 'UNVERIFIED',
    Pending: 'PENDING',
    Submitted: 'SUBMITTED',
    PositivelyVerified: 'POSITIVELY_VERIFIED',
    NegativelyVerified: 'NEGATIVELY_VERIFIED',
    Skipped: 'SKIPPED'
} as const;
export type CustomerPersonDTOLivenessCheckStatusEnum = typeof CustomerPersonDTOLivenessCheckStatusEnum[keyof typeof CustomerPersonDTOLivenessCheckStatusEnum];